Extract dmi_read/dmi_write into their own file
This allows them to be overridden without having to edit the source, and
for the prototypes to be included without pulling in all the
implementation details.
