+++ src/RVDM.c
@@ -29,22 +29,6 @@
 #include  "RVDM.h"
 
 // ================================================================
-// DMI interface (gdbstub invokes these functions)
-// These should be filled in with the appropriate mechanisms that
-// perform the actual DMI read/write on the RISC-V Debug module.
-
-void dmi_write (uint16_t addr, uint32_t data)
-{
-    fprintf (stderr, "RVDM.c: dmi_write(): Not yet implemented\n");
-}
-
-uint32_t  dmi_read  (uint16_t addr)
-{
-    fprintf (stderr, "RVDM.c: dmi_read(): Not yet implemented\n");
-    return 0;
-}
-
-// ================================================================
 // Debug Module address map
 
 // ----------------
