+++ src/RVDM.h
@@ -15,14 +15,6 @@
 //    Tue Oct 2 23:17:49 2018 -0700
 
 // ================================================================
-// DMI interface (gdbstub invokes these functions)
-// These should be filled in with the appropriate mechanisms that
-// perform the actual DMI read/write on the RISC-V Debug module.
-
-extern void      dmi_write (uint16_t addr, uint32_t data);
-extern uint32_t  dmi_read  (uint16_t addr);
-
-// ================================================================
 // Debug Module address map
 
 // ----------------
